requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
redis>=4.5.0
aiohttp>=3.8.0
facebook-business>=19.0.0
pandas>=2.0.0
pymongo>=4.6.0
motor>=3.3.0
apscheduler>=3.10.0
schedule>=1.2.0